% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_train.R
\name{dtrain}
\alias{dtrain}
\title{Density for the Train Delay Distribution}
\usage{
dtrain(x)
}
\arguments{
\item{x}{Numeric vector of quantiles.}
}
\value{
Numeric vector of densities.
}
\description{
Implements d(x) for Y with pdf f(y) = c(25 - y^2) on (-5, 5), c = 3/500.
}
\examples{
dtrain(0)
curve(dtrain, from = -10, to = 10)
}
\seealso{
\code{\link[=ptrain]{ptrain()}}, \code{\link[=qtrain]{qtrain()}}, \code{\link[=rtrain]{rtrain()}}

Other train-distribution: 
\code{\link{ptrain}()},
\code{\link{qtrain}()},
\code{\link{rtrain}()}
}
\concept{train-distribution}
