% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_train.R
\name{ptrain}
\alias{ptrain}
\title{CDF for the Train Delay Distribution}
\usage{
ptrain(q)
}
\arguments{
\item{q}{Numeric vector of quantiles.}
}
\value{
Numeric vector with F(q).
}
\description{
CDF for the Train Delay Distribution
}
\examples{
ptrain(0)
curve(ptrain, from = -10, to = 10)
}
\seealso{
Other train-distribution: 
\code{\link{dtrain}()},
\code{\link{qtrain}()},
\code{\link{rtrain}()}
}
\concept{train-distribution}
