% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_train.R
\name{qtrain}
\alias{qtrain}
\title{Quantile Function for the Train Delay Distribution}
\usage{
qtrain(p)
}
\arguments{
\item{p}{Numeric vector of probabilities in [0, 1].}
}
\value{
Numeric vector of quantiles.
}
\description{
Uses inverse transform computed in Python (robust bisection).
}
\examples{
qtrain(0.75)
qtrain(c(0.25, 0.5, 0.75))
}
\seealso{
Other train-distribution: 
\code{\link{dtrain}()},
\code{\link{ptrain}()},
\code{\link{rtrain}()}
}
\concept{train-distribution}
