% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr_train.R
\name{rtrain}
\alias{rtrain}
\title{Random Generation for the Train Delay Distribution}
\usage{
rtrain(n, seed = NULL)
}
\arguments{
\item{n}{Integer; number of observations to generate.}

\item{seed}{Optional integer seed for reproducibility (set in Python side).}
}
\value{
Numeric vector of length n.
}
\description{
Random Generation for the Train Delay Distribution
}
\examples{
set.seed(123)
x <- rtrain(10000)
hist(x, prob = TRUE, breaks = 60, col = 'gray')
curve(dtrain, add = TRUE, n = 501, lwd = 2, col = 2)
}
\seealso{
Other train-distribution: 
\code{\link{dtrain}()},
\code{\link{ptrain}()},
\code{\link{qtrain}()}
}
\concept{train-distribution}
